Add support for zip with prefixed bytes
Update archive/zip/reader.go to support zip archives that include
additional prefix bytes.

As zip file 'end of central directory record' structures are found by
searching backwards from the end of the file, it is actually possible
to construct an archive that contains additional bytes at the front.
This technique is occasionally used with Java 'jar' archives where a
bash script is embedded in the front of the archive to make it fully
executable.

Prior to this commit such archives could not be handled directly by go.